chore: split integration and peak picking
Also removed from the Chromatogram class and put it in the index root
