+++ src/index.js
@@ -1,3 +1,7 @@
 'use strict';
 
 exports.Chromatogram = require('./Chromatogram');
+
+// Chromatography utils
+exports.peakPicking = require('./peakPicking');
+exports.massInPeaks = require('./massInPeaks');
