+++ src/massInPeaks.js
@@ -1,52 +1,12 @@
 'use strict';
 
-const gsd = require('ml-gsd').gsd;
-
 /**
- * Calculate the array of peaks with their respective integrated MS
- * @param {Array<Number>} tic - TIC array of GC spectra
- * @param {Array<Number>} times - Times array of GC spectra
+ * Integrate MS spectra of a peak list
+ * @param {Array<Object>} peakList - List of GSD objects
  * @param {Array<Object>} sampleMS - MS array of GC spectra
- * @return {Array<Object>} - Peaks with their integrated MS
+ * @return {Array<Object>} - List of GSD objects with an extra 'ms' field with the integrated MS spectra
  */
-function massInPeaks(tic, times, sampleMS) {
-    // first peak selection
-    let peakList = gsd(times, tic, {
-        noiseLevel: 0,
-        realTopDetection: false,
-        smoothY: true,
-        sgOptions: {windowSize: 5, polynomial: 2},
-        heightFactor: 2,
-        boundaries: true
-    });
-
-    peakList.sort((a, b) => (a.right.index - a.left.index) - (b.right.index - b.left.index));
-    let medianDotsWidth = peakList[Math.floor((peakList.length - 1) / 2)];
-    medianDotsWidth = medianDotsWidth.right.index - medianDotsWidth.left.index;
-
-    // reset parameters
-    if (medianDotsWidth < 5) {
-        medianDotsWidth = 5;
-    }
-    if (medianDotsWidth % 2 === 0) {
-        medianDotsWidth -= 1;
-    }
-
-    // second peak selection
-    peakList = gsd(times, tic, {
-        noiseLevel: 0,
-        realTopDetection: false,
-        smoothY: true,
-        sgOptions: {windowSize: medianDotsWidth, polynomial: 2},
-        heightFactor: 2,
-        boundaries: true
-    });
-    peakList.sort((a, b) => a.height - b.height);
-
-    // filter height by factor
-    let medianHeight = peakList[Math.floor((peakList.length - 1) / 2)].height;
-    peakList = peakList.filter((val) => val.height > medianHeight * 2);
-
+function massInPeaks(peakList, sampleMS) {
     // integrate MS
     for (let i = 0; i < peakList.length; ++i) {
         let massDictionary = {};
